requirements.txt
Flask==1.1.2
fusionauth-client==1.18.8
